Refactor proposals on governance
